+++ jsplat-io-sog/src/main/java/de/javagl/jsplat/io/sog/meta/ShN.java
@@ -44,7 +44,7 @@ public class ShN
     /**
      * 256 floats; shared for all AC coefficients (§3.5)
      */
-    public float codebook[];
+    public double codebook[];
 
     /**
      * The files
